+++ .storybook/theme.js
@@ -1,4 +1,4 @@
-import { create } from '@storybook/theming/create';
+import { create } from 'storybook/theming/create';
 import ReagraphLogo from '../stories/assets/logo.svg';
 
 export default create({