+++ src/symbols/nodes/index.ts
@@ -3,6 +3,7 @@ export * from './Sphere';
 export * from './SphereWithIcon';
 export * from './Svg';
 export * from './SphereWithSvg';
+export * from './Badge';
 export * from './InstancedSpheres';
 export * from './InstancedText';
 export * from './InstancedIcon';